Enable SDI for Hello World Component and remove CDN cache headers
In this example, the Hello World component is configured for SDI in dispatcher layer. This configuration uses a nocache selector so the component is always delivered by publish server, while the rest of the page is cached in dispatcher. CDN cache-control headers have been removed so it does not cache the entire html request, and we can see SDI work in AEMaaCS.